Implemented support for SSM Parameter Storage
Previously the Travis build could not deploy the .env file, which contained the GitHub Access Token. This file was removed and the token is now located in SSM Parameter Storage. In the serverless.yml file the encrypted token will be provided for the cronjob as environment variable. The GitHub Organisation is also there, but as String.